refactor(entity): package model rename to entity
