--- server/src/main/java/pl/talkapp/server/Status.java
@@ -1,7 +1,0 @@
-package pl.talkapp.server;
-
-public enum Status {
-    ONLINE,
-    OFFLINE,
-    BUSY
-}
