+++ server/src/main/java/pl/talkapp/server/model/Status.java
@@ -1,40 +1,7 @@
 package pl.talkapp.server.model;
 
-import lombok.AccessLevel;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
-import lombok.experimental.FieldDefaults;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
-import java.util.List;
-
-@Entity
-@Table(name = "status")
-@Data
-@NoArgsConstructor
-@RequiredArgsConstructor
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class Status {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Long id;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false, unique = true)
-    @NonNull
-    pl.talkapp.server.Status name;
-
-    @OneToMany(mappedBy = "status")
-    List<User> users;
-
+public enum Status {
+    ONLINE,
+    OFFLINE,
+    BUSY
 }
