+++ server/src/main/java/pl/talkapp/server/repository/StatusRepository.java
@@ -2,7 +2,7 @@ package pl.talkapp.server.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import pl.talkapp.server.model.Status;
+import pl.talkapp.server.entity.Status;
 
 @Repository
 public interface StatusRepository extends JpaRepository<Status, Long> {
