+++ server/src/main/java/pl/talkapp/server/repository/UserRepository.java
@@ -2,7 +2,7 @@ package pl.talkapp.server.repository;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-import pl.talkapp.server.model.User;
+import pl.talkapp.server.entity.User;
 
 import java.util.Optional;
