+++ server/src/main/java/pl/talkapp/server/service/user/UserProfileService.java
@@ -1,6 +1,6 @@
 package pl.talkapp.server.service.user;
 
-import pl.talkapp.server.model.User;
+import pl.talkapp.server.entity.User;
 
 public interface UserProfileService {
     boolean changePassword(User user, String currentPassword, String newPassword);
