+++ server/src/main/java/pl/talkapp/server/service/user/UserServiceImpl.java
@@ -3,8 +3,8 @@ package pl.talkapp.server.service.user;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import pl.talkapp.server.model.Status;
-import pl.talkapp.server.model.User;
+import pl.talkapp.server.entity.Status;
+import pl.talkapp.server.entity.User;
 import pl.talkapp.server.repository.StatusRepository;
 import pl.talkapp.server.repository.UserRepository;
